Resolve remaining TODOs
