+++ src/packages/map/layers/group/GroupLayerCollection.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-import { getRecursiveLayers } from "../../model/getRecursiveLayers";
-import { LayerRetrievalOptions, RecursiveRetrievalOptions } from "../../shared";
+import { getRecursiveLayers } from "../shared/getRecursiveLayers";
+import { LayerRetrievalOptions, RecursiveRetrievalOptions } from "../shared/LayerRetrievalOptions";
 import { AbstractLayer } from "../AbstractLayer";
 import { GroupLayer } from "../GroupLayer";
 import { ChildrenCollection } from "../shared/ChildrenCollection";
