+++ src/packages/map/layers/shared/ChildrenCollection.ts
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-import { LayerRetrievalOptions } from "../../shared";
+import { LayerRetrievalOptions } from "./LayerRetrievalOptions";
 
 /**
  * Contains the children of a layer.
