+++ src/packages/map/layers/shared/SublayersCollection.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-import { getRecursiveLayers } from "../../model/getRecursiveLayers";
-import { LayerRetrievalOptions, RecursiveRetrievalOptions } from "../../shared";
+import { getRecursiveLayers } from "./getRecursiveLayers";
+import { LayerRetrievalOptions, RecursiveRetrievalOptions } from "./LayerRetrievalOptions";
 import { Sublayer } from "../unions";
 import { ChildrenCollection } from "./ChildrenCollection";
 import { GET_RAW_SUBLAYERS } from "./internals";
